+++ feature/transaction/src/main/java/dev/muffar/moneyfikasi/transaction/add_edit/AddEditTransactionViewModel.kt
@@ -74,7 +74,7 @@ class AddEditTransactionViewModel @Inject constructor(
                             state.copy(
                                 id = transactionId,
                                 type = type,
-                                amount = amount.toLong().formatThousand(),
+                                amount = amount.formatThousand(),
                                 category = category,
                                 wallet = wallet,
                                 note = note ?: "",
