+++ shared/common-ui/src/main/java/dev/muffar/moneyfikasi/common_ui/component/text_input/WalletPicker.kt
@@ -103,7 +103,7 @@ fun WalletPicker(
                         Column {
                             Text(text = wallet.name)
                             Text(
-                                text = wallet.balance.toLong().formatThousand(),
+                                text = wallet.balance.formatThousand(),
                                 style = MaterialTheme.typography.titleMedium.copy(
                                     fontWeight = FontWeight.Medium
                                 )
